Added additional interface for Grades Service layer and changed final_grade in Grades model
