+++ src/main/java/com/faculty/managementsystem/controller/GradesController.java
@@ -1,7 +1,7 @@
 package com.faculty.managementsystem.controller;
 
 import com.faculty.managementsystem.model.Grades;
-import com.faculty.managementsystem.service.GradesService;
+import com.faculty.managementsystem.service.GradesServiceImp;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,40 +12,40 @@ import java.util.Optional;
 @RequestMapping("/grades")
 @AllArgsConstructor
 public class GradesController {
-    private final GradesService gradesService;
+    private final GradesServiceImp gradesServiceImp;
 
     @GetMapping("/all")
     public List<Grades> getAllGrades() {
-        return gradesService.getAllGrades();
+        return gradesServiceImp.getAllGrades();
     }
 
     @GetMapping("/all/{id}")
     public Optional<Grades> getAllGradesById(@PathVariable Long id) {
-        return gradesService.getAllGradesById(id);
+        return gradesServiceImp.getAllGradesById(id);
     }
 
 
     @PostMapping("/add")
     public String addGrades(@RequestBody Grades grades) {
-        gradesService.addGrades(grades);
+        gradesServiceImp.addGrades(grades);
         return "Grades were successfully saved!";
     }
 
     @PutMapping("/update/{id}")
     public String updateGrades(@PathVariable Long id, @RequestBody Grades grades) {
-        gradesService.updateGrades(id, grades);
+        gradesServiceImp.updateGrades(id, grades);
         return String.format("The grades at id: %s were updated!", id);
     }
 
     @DeleteMapping("/deleteAll")
     public String deleteAllGrades() {
-        gradesService.deleteAllGrades();
+        gradesServiceImp.deleteAllGrades();
         return "All of the grades were deleted!";
     }
 
     @DeleteMapping("/delete/{id}")
     public String deleteGrades(@PathVariable Long id) {
-        gradesService.deleteGradesById(id);
+        gradesServiceImp.deleteGradesById(id);
         return String.format("The grades at id: %s were deleted!", id);
     }
